Brea down about page to components.
